requirements.txt
discord==2.1.0
python-telegram-bot==13.15